requirements.txt
watchdog
firebase_admin
websockets
 websocket-client